nit: rename mockDialogContext and move to testing folder
